requirements.txt
streamlit
mysql-connector-python
python-dotenv
Pillow
tensorflow
pandas
numpy
requests
pycountry
plotly